More tree fixes
